fix: rename memo -> note + add bip39 mnemonic check
