+++ src/containers/Wallet/Proposals/Modal/Memo.js
@@ -1,5 +1,5 @@
 import * as PropTypes from 'prop-types';
-import { ValidateMemo } from './_validation';
+import { ValidateMemo } from '../../../common/_validation';
 import { connect } from 'react-redux';
 import { setTxVoteMemo } from '../../../../actions/transactions/vote';
 import React from 'react';
@@ -19,8 +19,8 @@ const Memo = (props) => {
         <TextArea
             className="form-control seed-text-field"
             error={props.input.error}
-            name="Memo"
-            placeholder="Enter Memo"
+            name="Note"
+            placeholder="Enter Note"
             required={true}
             rows={3}
             value={props.input.value}
