+++ src/containers/Wallet/Proposals/Modal/index.js
@@ -32,7 +32,7 @@ const Modal = ({
                 <div className="form-group">
                     <Label
                         className=""
-                        label="Memo"
+                        label="Note"
                     />
                     <Memo/>
                 </div>
