+++ src/containers/Wallet/Send/Modal/index.js
@@ -48,7 +48,7 @@ const Modal = ({
                 <div className="form-group">
                     <Label
                         className=""
-                        label="Memo"
+                        label="Note"
                     />
                     <Memo/>
                 </div>
