+++ src/containers/Wallet/Validators/ModalDelegate/index.js
@@ -46,7 +46,7 @@ const ModalDelegate = (props) => {
                 <div className="form-group">
                     <Label
                         className=""
-                        label="Memo"
+                        label="Note"
                     />
                     <Memo/>
                 </div>
