+++ src/containers/Wallet/Validators/ModalRedelegate/index.js
@@ -54,7 +54,7 @@ const ModalRedelegate = (props) => {
                 <div className="form-group">
                     <Label
                         className=""
-                        label="Memo"
+                        label="Note"
                     />
                     <Memo/>
                 </div>
