+++ src/containers/Wallet/WithdrawRewards/Modal/index.js
@@ -40,7 +40,7 @@ const Modal = (props) => {
                 <div className="form-group">
                     <Label
                         className=""
-                        label="Memo"
+                        label="Note"
                     />
                     <Memo/>
                 </div>
