+++ src/containers/common/_validation.js
@@ -1,4 +1,12 @@
+const bip39 = require('bip39');
+
 export const ValidateMemo = (value) => {
+    if (bip39.validateMnemonic(value.trim())) {
+        return new Error(
+            'This looks like a mnemonic. Please do not attach the mnemonic ' +
+            'or other sensitive data to note, as it is public and can be seen by anyone.');
+    }
+
     return new Error('');
 };
